+++ app/client/src/components/ui/MarqueeSingleValue.js
