+++ app/client/src/components/ui/SliderWrapper.js
