+++ app/client/src/services/ReleaseService.js
@@ -1,9 +1,6 @@
 import Api from './Api'
 
 const service = {
-  getReleaseNotes() {
-    return Api().get('/api/release-notes')
-  },
   getReleases(offset = 0, limit = 3) {
     return Api().get('/api/releases', { params: { offset, limit } })
   },
