+++ app/client/src/services/index.js
@@ -7,3 +7,4 @@ export { default as StatsService } from './StatsService'
 export { default as GameService } from './GameService'
 export { default as ReleaseService } from './ReleaseService'
 export { default as adminSSE } from './AdminSSE'
+export { default as WarningService } from './WarningService'
