+++ app/client/src/views/Dashboard.js
@@ -23,15 +23,15 @@ import CheckIcon from '@mui/icons-material/Check'
 import LinkIcon from '@mui/icons-material/Link'
 import LocalOfferIcon from '@mui/icons-material/LocalOffer'
 import VideoCards from '../components/cards/VideoCards'
-import LoadingSpinner from '../components/misc/LoadingSpinner'
+import LoadingSpinner from '../components/ui/LoadingSpinner'
 import { VideoService, GameService, TagService } from '../services'
 import Select from 'react-select'
 import SnackbarAlert from '../components/alert/SnackbarAlert'
-import TagChip from '../components/misc/TagChip'
+import TagChip from '../components/ui/TagChip'
 
-import selectFolderTheme from '../common/reactSelectFolderTheme'
-import OutlinedIconButton from '../components/misc/OutlinedIconButton'
-import MarqueeSingleValue from '../components/misc/MarqueeSingleValue'
+import { folderSelectTheme as selectFolderTheme } from '../common/reactSelectThemes'
+import OutlinedIconButton from '../components/ui/OutlinedIconButton'
+import MarqueeSingleValue from '../components/ui/MarqueeSingleValue'
 import { SORT_OPTIONS } from '../common/constants'
 import { inputSx, dialogPaperSx, dialogTitleSx } from '../common/modalStyles'
