+++ app/client/src/views/ImageFeed.js
@@ -21,11 +21,11 @@ import Select from 'react-select'
 import ImageCards from '../components/cards/ImageCards'
 import EditImageModal from '../components/modal/EditImageModal'
 import { ImageService } from '../services'
-import LoadingSpinner from '../components/misc/LoadingSpinner'
+import LoadingSpinner from '../components/ui/LoadingSpinner'
 import SnackbarAlert from '../components/alert/SnackbarAlert'
-import selectFolderTheme from '../common/reactSelectFolderTheme'
-import OutlinedIconButton from '../components/misc/OutlinedIconButton'
-import MarqueeSingleValue from '../components/misc/MarqueeSingleValue'
+import { folderSelectTheme as selectFolderTheme } from '../common/reactSelectThemes'
+import OutlinedIconButton from '../components/ui/OutlinedIconButton'
+import MarqueeSingleValue from '../components/ui/MarqueeSingleValue'
 import { SORT_OPTIONS } from '../common/constants'
 
 const ImageFeed = ({ authenticated, searchText, cardSize, selectedImageFolder, onImageFoldersLoaded, onImageFolderChange, showFolderDropdown, uploadTick }) => {
