+++ app/server/requirements.txt
@@ -6,7 +6,6 @@ Flask-Login==0.6.3
 Flask-Migrate==4.1.0
 Flask-SQLAlchemy==3.1.1
 Flask-WTF==1.2.2
-future==0.18.3
 greenlet>=3.1.0
 gunicorn==23.0.0
 itsdangerous==2.2.0
